+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/validation/model/Rule.kt
@@ -1,0 +1,18 @@
+package com.r0adkll.deckbuilder.arch.domain.features.validation.model
+
+
+import android.support.annotation.StringRes
+import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
+
+
+interface Rule {
+
+    /**
+     * Check if the list of cards, the current representation of a deck, is valid for this implementation of
+     * the rule
+     *
+     * @param cards the list of cards in a deck to check/validate
+     * @return the error message, as a [StringRes], or null if cards are valid for this rule
+     */
+    @StringRes fun check(cards: List<PokemonCard>): Int?
+}