+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/importer/DeckImportPresenter.kt
@@ -5,6 +5,7 @@ import com.r0adkll.deckbuilder.arch.domain.features.ptcgo.repository.PTCGOConver
 import com.r0adkll.deckbuilder.arch.ui.components.presenter.Presenter
 import com.r0adkll.deckbuilder.arch.ui.features.importer.DeckImportUi.State
 import com.r0adkll.deckbuilder.arch.ui.features.importer.DeckImportUi.State.Change
+import com.r0adkll.deckbuilder.util.extensions.logState
 import com.r0adkll.deckbuilder.util.extensions.plusAssign
 import timber.log.Timber
 import javax.inject.Inject
@@ -28,7 +29,7 @@ class DeckImportPresenter @Inject constructor(
                 .doOnNext { Timber.d(it.logText) }
 
         disposables += conversion.scan(ui.state, State::reduce)
-                .doOnNext { state -> Timber.v("    --- $state") }
+                .logState()
                 .subscribe { ui.render(it) }
 
     }