+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/search/pageadapter/ResultsPagerAdapter.kt
@@ -11,12 +11,14 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.animation.*
 import android.view.animation.Animation.RELATIVE_TO_SELF
+import com.ftinc.kit.kotlin.extensions.color
 import com.ftinc.kit.kotlin.extensions.dpToPx
 import com.ftinc.kit.widget.EmptyView
 import com.jakewharton.rxrelay2.Relay
 import com.r0adkll.deckbuilder.R
 import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
 import com.r0adkll.deckbuilder.arch.ui.features.deckbuilder.EditCardIntentions
+import com.r0adkll.deckbuilder.arch.ui.features.missingcards.MissingCardsActivity
 import com.r0adkll.deckbuilder.arch.ui.features.search.adapter.SearchResultsRecyclerAdapter
 import com.r0adkll.deckbuilder.arch.ui.widgets.PokemonCardView
 import io.pokemontcg.model.SuperType
@@ -36,7 +38,7 @@ class ResultsPagerAdapter(
 
     override fun instantiateItem(container: ViewGroup, position: Int): Any {
         val view = inflater.inflate(R.layout.layout_deck_supertype, container, false)
-        val vh = SearchResultViewHolder(view, scrollHideListener, pokemonCardLongClicks, editCardIntentions)
+        val vh = SearchResultViewHolder(view, position, scrollHideListener, pokemonCardLongClicks, editCardIntentions)
         view.tag = vh
         viewHolders[position] = vh
 
@@ -89,6 +91,24 @@ class ResultsPagerAdapter(
     }
 
 
+    fun showEmptyResults(type: SuperType) {
+        when(type) {
+            SuperType.POKEMON -> viewHolders[0]?.showEmptyResults()
+            SuperType.TRAINER -> viewHolders[1]?.showEmptyResults()
+            SuperType.ENERGY -> viewHolders[2]?.showEmptyResults()
+        }
+    }
+
+
+    fun showEmptyDefault(type: SuperType) {
+        when(type) {
+            SuperType.POKEMON -> viewHolders[0]?.showEmptyDefault()
+            SuperType.TRAINER -> viewHolders[1]?.showEmptyDefault()
+            SuperType.ENERGY -> viewHolders[2]?.showEmptyDefault()
+        }
+    }
+
+
     fun showError(type: SuperType, description: String) {
         when(type) {
             SuperType.POKEMON -> viewHolders[0]?.showError(description)
@@ -117,7 +137,8 @@ class ResultsPagerAdapter(
 
 
     private class SearchResultViewHolder(
-            itemView: View,
+            val itemView: View,
+            val position: Int,
             scrollHideListener: KeyboardScrollHideListener,
             pokemonCardLongClicks: Relay<PokemonCardView>,
             editCardIntentions: EditCardIntentions
@@ -130,6 +151,11 @@ class ResultsPagerAdapter(
 
         init {
             emptyView.setIcon(R.drawable.ic_empty_search)
+            emptyView.setEmptyMessage(when(position) {
+                0 -> R.string.empty_search_pokemon_message
+                1 -> R.string.empty_search_trainer_message
+                else -> R.string.empty_search_energy_message
+            })
 
             adapter.setEmptyView(emptyView)
             adapter.setOnItemClickListener { editCardIntentions.addCardClicks.accept(listOf(it)) }
@@ -163,13 +189,37 @@ class ResultsPagerAdapter(
         }
 
 
+        fun showEmptyResults() {
+            emptyView.setEmptyMessage(when(position) {
+                0 -> R.string.empty_search_results_pokemon_message
+                1 -> R.string.empty_search_results_trainer_message
+                else -> R.string.empty_search_results_energy_message
+            })
+            emptyView.setActionLabelRes(R.string.empty_search_missing_card)
+            emptyView.actionColor = emptyView.color(R.color.red_500)
+            emptyView.setOnActionClickListener {
+                MissingCardsActivity.show(itemView.context)
+            }
+        }
+
+
+        fun showEmptyDefault() {
+            emptyView.setEmptyMessage(when(position) {
+                0 -> R.string.empty_search_pokemon_message
+                1 -> R.string.empty_search_trainer_message
+                else -> R.string.empty_search_energy_message
+            })
+            emptyView.actionLabel = null
+        }
+
+
         fun showError(description: String) {
             emptyView.emptyMessage = description
         }
 
 
         fun hideError() {
-            emptyView.setEmptyMessage(R.string.empty_search_category)
+            showEmptyDefault()
         }