+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/unifiedsearch/SearchUi.kt
@@ -27,6 +27,8 @@ interface SearchUi : StateRenderer<SearchUi.State> {
         fun showFilterEmpty(enabled: Boolean)
         fun setQueryText(text: String)
         fun setResults(cards: List<PokemonCard>)
+        fun showEmptyResults()
+        fun showEmptyDefault()
         fun showLoading(isLoading: Boolean)
         fun showError(description: String)
         fun hideError()
@@ -48,7 +50,7 @@ interface SearchUi : StateRenderer<SearchUi.State> {
             is Change.Error -> this.copy(error = change.description, isLoading = false)
             is Change.QuerySubmitted -> this.copy(query = change.query)
             is Change.FilterChanged -> this.copy(filter = change.filter,
-                    results = if (change.filter.isEmpty && query.isBlank()) emptyList() else results)
+                    results = if (change.filter.isEmptyWithoutField && query.isBlank()) emptyList() else results)
             is Change.ResultsLoaded -> this.copy(results = change.results, error = null, isLoading = false)
         }