+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/widgets/QuickTipView.kt
@@ -94,13 +94,11 @@ class QuickTipView @JvmOverloads constructor(
 
 
     override fun onTouchEvent(event: MotionEvent): Boolean {
-
         when (event.action) {
             MotionEvent.ACTION_UP -> {
                 view?.let { hide(it) }
             }
         }
-
         return true
     }
 
@@ -108,17 +106,15 @@ class QuickTipView @JvmOverloads constructor(
     override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
         super.onSizeChanged(w, h, oldw, oldh)
         if (w != oldw || h != oldh) {
-            val radius = getRadius().toInt() * 2
             bitmapCache?.recycle()
-            bitmapCache = Bitmap.createBitmap(radius, radius, Bitmap.Config.ARGB_8888)
-            canvasCache = Canvas(bitmapCache)
+            bitmapCache = null
         }
     }
 
 
     override fun onDraw(canvas: Canvas?) {
         if (bitmapCache == null) {
-            val radius = getRadius().toInt()
+            val radius = getRadius().toInt() * 2
             bitmapCache?.recycle()
             bitmapCache = Bitmap.createBitmap(radius, radius, Bitmap.Config.ARGB_8888)
             canvasCache = Canvas(bitmapCache)