+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-# 1.0.0-RC
+# 1.2.0
+* Missing card feedback form
+* Disabled analytics during robo-tests
+* Added missing card prompt when search returns no results
+
+# 1.1.1
+* Fixed bug with quickstart menu rendering incorrectly
+* Added improved search functionality to allow the user to search other fields of the card (i.e. text, ability name, attack text, etc)
+* Minor UI improvements
+
+# 1.1.0
+* Improved the builder interface to show the non-evolution line Pokémons in a grid pattern similar to trainer and energy cards
+* Improved UI in the card detail screen to allow you to add more copies of that card, or of variants and evolved forms
+* Removed hard limits on deck rules in favor of soft warnings.
+* Added basic Standard/Expanded format indicators in deck builder interface
+* Minor bug fixes
+
+# 1.0.0
 * Fixed Quickstart view on tablets
 * Fixed search when just typing 'n' or 'N' to wrap in quotes
 * Fixed issue where '-' and '+' buttons weren't showing in search results